// Desktop/UNEXT_LMS_PORTAL/src/app/enums/rubricOperations.ts
export enum RubricOperations{
    ACTIVE = 'active',
    ACTIVE_STATE = 'active-state',
    ADD_CRITERIA_ABOVE = 'addCriteriaAbove',
    ADD_CRITERIA_BELOW = 'addCriteriaBelow',
    ADD_LEVEL_LEFT = 'addLevelLeft',
    ADD_LEVEL_RIGHT = 'addLevelRight',
    ATTACH_RUBRIC = 'attachRubric',
    CANCEL = 'cancel',
    CLOSE = 'close',
    CLICK = 'click',
    COPY = 'copy',
    COPY_CRITERIA = 'copyCriteria',
    COURSE_LEVEL = 'CourseLevel',
    COURSE = 'course',
    DELETE = 'delete',
    DRAFT = "draft",
    DRAFT_STATE = 'draft-state',
    CREATE = 'create',
    CREATE_RUBRIC = 'Create Rubric',
    CRITERIA = 'criteria',
    EDIT = 'edit',
    EDIT_RUBRIC = 'Edit Rubric',
    GRADED_LEVEL = 'gradedLevel',
    IN_USE="inUse",
    LEVELS = 'levels',
    PERCENTAGE = 'percentage',
    PROGRAM = 'program',
    PROGRAM_LEVEL = 'ProgramLevel',
    REMOVE_CRITERIA = 'removeCriteria',
    REMOVE_LEVEL = 'removeLevel',
    RUBRIC_SELECTION = 'rubricSelection',
    SAVE_CRITERIA = 'saveCriteria',
    SAVE_AS_DRAFT = 'saveAsDraft',
    SELECTION = 'selection',
    SUCCESS = "success",
    UPDATE = 'update',
    USE_STATE = 'use-state',
    VALUE_CHANGES = 'valueChanges',
    WEIGHTAGE = 'weightage',
    RUBRIC_DUPLICATE_TITLE_ERROR_MESSAGE = 'Rubric title already exists, please create a different one'
}