// Desktop/UNEXT_LMS_PORTAL/src/app/enums/service.ts
export enum Service {
  CONFIGURATION = 'configservice',
  MAGE_ASSESSMENT_SERVICE = 'assessmentservice_mage',
  COMMON_SERVICE = 'commonservice',
  CONTENT_SERVICE = 'contentservice',
  COURSE_SERVICE = 'courseservice',
  ELASTICSEARCH_SERVICE = 'elasticsearchservice',
  FEEDBACK_SERVICE = 'feedbackservice',
  FORMS_SERVICE = 'formsservice',
  ORGANIZATION_SERVICE = 'organizationservice',
  PROGRAM_SERVICE = 'programservice',
  COURSEBOOK_SERVICE = 'coursebookservice',
  INSIGHTS_SERVICE = 'insightsservice',
  TELEMETRY_SERVICE = 'telemetryservice',
  USER_SERVICE = 'userservice',
  AUTHORIZATION_SERVICE = 'authorizationservice',
  RESULT_SERVICE = 'resultservice',
  ASSESSMENT_SERVICE = 'assessmentservice',
  BEHAVIOURAL_REPORT_SERVICE = 'behaviouralreportservice',
  QUESTION_SERVICE = 'questionservice',
  CATALOGUE_SERVICE = 'catalogueservice',
  USERGROUP_SERVICE = 'usergroupservice',
  TIMER_SERVICE = 'timerservice',
  TIMER_BATCH = 'timerbatch',
  USER_CREATION_BATCH = 'usercreationbatch',
  USER_STATUS_MANAGEMENT_BATCH = 'userstatusmanagementbatch',
  LMS_BATCH_SERVICE = 'lmsbatchservice',
  NOTIFICATION_SERVICE = 'notificationservice',
  SCHEDULING_SERVICE = 'schedulingservice',
  PLAGIARISM_CHECK_SERVICE = 'plagiarismcheckservice',
  COURSE_LIST_SERVICE= 'courselistservice',
  
  RPSESSION_SERVICE = 'rpsessionservice',
  RPLICENSE_SERVICE = 'rplicesenseservice',
  RPRESULT_SERVICE = 'rpresultservice',
  RPSIGNALING_SERVICE = 'rpsignalingservice',
  RPCONFIGURATION_SERVICE = 'rpconfigurationservice',
  RPREPORT_SERVICE = 'rpreportservice',
  RP_VIDEO_MANAGER_SERVICE = 'rpvideomanagerservice',
  RP_VIDEO_FORMAT_BATCH_SERVICE = 'rpvideoformatbatchservice',
  RP_MEDIA_ARCHIVAL = 'rpmediaarchival',

  PE_METADATA_SERVICE = 'pemetadataservice',
  TIMER_SERVICE_V2 = 'timerservicev2',

  GOALS_SERVICE = 'goalsservice',

  CONTENT_BUILDER_SERVICE = 'contentbuilderservice',

  DUMMY = 'dummyservice',
  DUMMY1 = 'dummyservice1',
  DUMMY2 = 'dummyservice2',
  DUMMY3 = 'dummyservice3',
  DUMMY4 = 'dummyservice4',
}